requirements.txt
openai
requests
python-dotenv
tk
pytest
pytest-cov
